pyproject.toml
[project]
name = "repomind"
version = "0.1.0"
description = "Context-aware code change assistant"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.0",
    "gitpython>=3.1.43",
    "mcp>=1.0.0",
    "openai>=1.30.0",
    "python-dotenv>=1.0.1",
    "sentence-transformers>=3.0.0",
    "streamlit>=1.42.0",
    "tree-sitter>=0.25.0",
    "tree-sitter-go>=0.23.4",
    "tree-sitter-javascript>=0.23.1",
    "tree-sitter-python>=0.23.6",
    "tree-sitter-typescript>=0.23.2",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
